feat(meet): share header with preview
